Refactor AddIncident component to use IncidentForm
The AddIncident component was heavily refactored to delegate most of its functionality to a new IncidentForm object. This simplifies the AddIncident component, improving code clarity and maintainability and leaving the component to mostly handle events and authorization.
